Introducing validators
It is better to provide user a summary of what's wrong, than to force
him to iterate over errors one at a time.

There are two correct way how to setup validators:
    1) in a method annotated with @Setup (newly added)
    2) from the outside

Please note that there are two incorrect ways of doing so:
    1) in the @BeforeParse method - that would add validators before
       every parsing
    2) in a constructor of the Options subclass - that introduces
       dependency on ArgumentParser, and adds those validators before
       options are added

The biggest change introduced is that many exceptions thrown by the
librery are now checked. Though the user now needs to wrap his code by
the same boilerplate, it forces him to make the error output more clean.
Otherwise it included stack trace from the library, which is not
desirable.